fix prettier drift, move auth method-view model out of component file
